[Win Python] REGRESSION(r284398): run-api-tests: FileNotFoundError raised: [WinError 2] The system cannot find the file specified https://bugs.webkit.org/show_bug.cgi?id=232037
Reviewed by Jonathan Bedard.

r284398 temporarily copied TestWTF program to ToBeListed to list
test cases. However, on Windows, TestWTF.exe is just a wrapper
program loading TestWTFLib.dll. If I just rename TestWTF.exe to
ToBeListed.exe, ToBeListed.exe tries to load ToBeListedLib.dll.

* Scripts/webkitpy/api_tests/manager.py:
(Manager._collect_tests): Do not copy the test program to
ToBeListed on Windows.

Canonical link: https://commits.webkit.org/243357@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@284637 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Tools/Scripts/webkitpy/api_tests/manager.py
@@ -106,10 +106,12 @@ class Manager(object):
             if canonicalized_binary not in specified_binaries:
                 continue
 
-            to_be_listed = self.host.filesystem.join(self.host.filesystem.dirname(path), 'ToBeListed')
-            try:
+            to_be_listed = path
+            if not self._port.host.platform.is_win():
+                to_be_listed = self.host.filesystem.join(self.host.filesystem.dirname(path), 'ToBeListed')
                 self.host.filesystem.copyfile(path, to_be_listed)
                 self.host.filesystem.copymode(path, to_be_listed)
+            try:
                 output = self.host.executive.run_command(
                     Runner.command_for_port(self._port, [to_be_listed, '--gtest_list_tests']),
                     env=self._port.environment_for_api_tests())
@@ -118,7 +120,8 @@ class Manager(object):
                 _log.error('Failed to list {} tests'.format(canonicalized_binary))
                 raise
             finally:
-                self.host.filesystem.remove(to_be_listed)
+                if not self._port.host.platform.is_win():
+                    self.host.filesystem.remove(to_be_listed)
 
         if len(args) == 0:
             return sorted(available_tests)
